Added navigation bar: dropdown menu
