--- app/content.php
@@ -1,24 +1,0 @@
-<link rel="stylesheet" href="CSS/all.css" type="text/css">
-<?php
-	function nav(){
-		echo "<nav><ul>
-				<li><a href=\"home.php\">HOME</a></li>
-				<li><a href=\"classes.php\">CLASSES</a></li>
-				<li><a href=\"grades.php\">GRADES</a></li>
-				<li><a href=\"resources.php\">RESOURCES</a></li>
-				<li><a href=\"contact.php\">CONTACT US</a></li>
-			</ul></nav>";
-		}
-
-	function head(){
-		echo "<header>
-				<h1>SCHOOL COOL TRACKER</h1>
-			</header>";
-		}
-
-	function foot(){
-		echo"<footer>
-				<h3>Stay cool.</h3>
-			</footer>";
-		}
-?>
