+++ app/content/footer.php
@@ -1,0 +1,8 @@
+<link rel="stylesheet" href="CSS/all.css" type="text/css">
+<?php
+	function foot(){
+		echo "<footer>
+				<h3>Stay cool.</h3>
+			</footer>";
+		}
+?>
