+++ app/content/header.php
@@ -1,0 +1,8 @@
+<link rel="stylesheet" href="CSS/all.css" type="text/css">
+<?php
+	function head(){
+		echo "<header>
+				<h1>SCHOOL COOL TRACKER</h1>
+			</header>";
+		}
+?>
