+++ app/css/all.css
@@ -26,15 +26,18 @@ header{
     padding-left: 15px 15px 0;
 }
 
+/*   ++++               NAV BAR            ++++        */
+
 nav{
     background: #06bbff;
     height:50px;
     line-height: 75px;
     opacity: .73;
+    overflow: hidden;
 }
 
-nav li:nth-child(1), nav li:nth-child(2),
-nav li:nth-child(3), nav li:nth-child(4) {
+nav li:nth-child(1), nav li:nth-child(3),
+nav li:nth-child(4), nav li:nth-child(5) {
     background: #57baff;
 }
 
@@ -53,13 +56,58 @@ nav a{
     line-height: 50px;
     display: block;
     color: white;
+    font-size: 16px;
 }
 
-nav a:hover{
+nav a:hover, .dropdown:hover .dropbtn{
     background: #aafef9;
     color: inherit;
     text-decoration: overline;
 }
+
+.dropdown {
+    overflow: hidden;
+    background-color: #57baff;
+    text-align: center;
+    padding-left: 40px;
+}
+
+.dropdown .dropbtn {
+    font-size: 16px;
+    border: none;
+    outline: none;
+    color: white;
+    font-family: inherit;
+    float: left;
+    width: 20%;
+    background-color: #57baff;
+}
+
+.dropdown-content {
+    display: none;
+    position: absolute;
+    background-color: inherit;
+    max-width: 160px;
+    box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
+    z-index: 1;
+}
+
+.dropdown-content a {
+    float: none;
+    color: white;
+    padding: 12px 16px;
+    display: block;
+    text-align: center;
+}
+
+.dropdown-content a:hover {
+    background-color: #aafef9;
+}
+
+.dropdown:hover .dropdown-content {
+    display: block;
+}
+/* end nav */
 
 main{
     background: #c1e5f9;
