+++ app/grades.php
@@ -1,38 +1,38 @@
 <?php
-	require('content.php')
+	require('/content/content.php')
 ?>
 
 <html>
     <head>
-        <title>School Cool gradesx</title>
+        <title>School Cool :: Grades</title>
     </head>
     <body>
-        
+
 		<?php
-			head();			
+			head();
 		?>
-		
+
 
         <div id="wrapper">
 			<?php
 				nav();
 			?>
 
-            
+
 
             <main>
 
             </main>
 
             <aside>
-			
+
 			</aside>
 
 			<?php
-			foot();			
+			foot();
 		?>
 
-            
+
         </div>
     </body>
 </html>
